order corrections

+++ app/Http/Controllers/HotelController.php
@@ -24,7 +24,7 @@ class HotelController extends Controller
             'price-desc' => Hotel::orderBy('price', 'desc')->get(),
             default => Hotel::all()
         };
-        
+
         $countries = Country::all();
         // $hotels = Hotel::all();
         // $countries = Country::where('id', '=', $countryId)->first();  
@@ -54,11 +54,11 @@ class HotelController extends Controller
             $request->all(),
             [
                 'hotel_name' => ['required', 'min:5', 'max:64'],
-                'hotel_price' => ['required', 'min:5', 'max:64'],
+                'hotel_price' => ['required', 'min:2', 'max:64'],
                 'hotel_trip_time' => ['required', 'min:5', 'max:64'],
             ],
         );
-        
+
         if ($validator->fails()) {
             $request->flash();
             return redirect()->back()->withErrors($validator);
@@ -68,7 +68,7 @@ class HotelController extends Controller
         $hotel->name = $request->hotel_name;
         $hotel->price = $request->hotel_price;
         $hotel->trip_time = $request->hotel_trip_time;
-        
+
         // ========================== Photo file ==========================
 
         if ($request->file('hotel_photo')) {
@@ -77,8 +77,8 @@ class HotelController extends Controller
             $ext = $photo->getClientOriginalExtension();  //get extention of the file
             $name = pathinfo($photo->getClientOriginalName(), PATHINFO_FILENAME); //original file name
 
-            $file = $name. '-' . rand(100, 111). '.' . $ext;  //create new name for the file
-            $photo->move(public_path().'/images', $file); //move file from tmp
+            $file = $name . '-' . rand(100, 111) . '.' . $ext;  //create new name for the file
+            $photo->move(public_path() . '/images', $file); //move file from tmp
 
             $hotel->photo = asset('/images') . '/' . $file; //read file path as url
         }
@@ -124,7 +124,31 @@ class HotelController extends Controller
         $hotel->name = $request->hotel_name;
         $hotel->price = $request->hotel_price;
         $hotel->trip_time = $request->hotel_trip_time;
-        $hotel->photo = $request->hotel_photo;
+        // $hotel->photo = $request->hotel_photo;
+
+        // ========================== Photo file ==========================
+
+        if ($request->file('hotel_photo')) {
+
+            $name = pathinfo($hotel->photo, PATHINFO_FILENAME);
+            $ext = pathinfo($hotel->photo, PATHINFO_EXTENSION);
+
+            $path = asset('/images') . '/' . $name . '.' . $ext;
+
+            if (file_exists($path)) {
+                unlink($path);
+            }
+
+            $photo = $request->file('hotel_photo');
+            $ext = $photo->getClientOriginalExtension();  //get extention of the file
+            $name = pathinfo($photo->getClientOriginalName(), PATHINFO_FILENAME); //original file name
+
+            $file = $name . '-' . rand(100, 999) . '.' . $ext;  //create new name for the file
+            $photo->move(public_path() . '/images', $file); //move file from tmp
+
+            $hotel->photo = asset('/images') . '/' . $file; //read file path as url
+        }
+
         $hotel->country_id = $request->country_id;
         $hotel->save();
         return redirect()->route('hotel.index')->with('pop_message', 'Successfully edited!');
@@ -138,21 +162,36 @@ class HotelController extends Controller
      */
     public function destroy(Hotel $hotel)
     {
-        $hotel->delete();
-        return redirect()->route('hotel.index')->with('pop_message', 'Successfully deleted!');
+        if ($hotel->photo) {
+
+            $name = pathinfo($hotel->photo, PATHINFO_FILENAME);
+            $ext = pathinfo($hotel->photo, PATHINFO_EXTENSION);
+
+            $path = public_path() . '/images/' . $name . '.' . $ext;
+
+            if (file_exists($path)) {
+                unlink($path);
+            }
+        }
+
+        if (!$hotel->orders->count()) {
+            $hotel->delete();
+            return redirect()->route('hotel.index')->with('pop_message', 'Successfully deleted!');
+        }
+        return redirect()->back()->with('pop_message', 'This hotel has orders and can not be deleted!');
     }
 
-    public function deletePicture(Hotel $hotel) 
+    public function deletePicture(Hotel $hotel)
     {
         $name = pathinfo($hotel->photo, PATHINFO_FILENAME);
         $ext = pathinfo($hotel->photo, PATHINFO_EXTENSION);
 
-        $path = asset('/images') . '/' . $name . '.' . $ext;
+        $path = public_path() . '/images/' . $name . '.' . $ext;
 
-        if(file_exists($path)) {
+        if (file_exists($path)) {
             unlink($path);
         }
-        
+
         $hotel->photo = null;
         $hotel->save();
